[M3] Restructure the solution in accordance with the feedback
