+++ Epam.ReportPortal.Automation/Epam.ReportPortal.Automation.ApiBusinessLayer/ApiSteps/BaseApiSteps.cs
@@ -1,0 +1,17 @@
+﻿using NUnit.Framework;
+
+namespace Epam.ReportPortal.Automation.ApiBusinessLayer.ApiSteps
+{
+    public abstract class BaseApiSteps
+    {
+        protected BaseApiSteps()
+        {
+            // TODO: for example init rest with user/password
+        }
+
+        public void ValidateResponseIsOk()
+        {
+            Assert.Fail("TODO: Not implemented");
+        }
+    }
+}
