+++ Epam.ReportPortal.Automation/Epam.ReportPortal.Automation.ApiBusinessLayer/ApiSteps/Dashboards/DashboardsApiSteps.cs
@@ -1,0 +1,15 @@
+﻿namespace Epam.ReportPortal.Automation.ApiBusinessLayer.ApiSteps.Dashboards;
+
+public class DashboardsApiSteps : BaseApiSteps
+{
+    public List<string> GetDashboards()
+    {
+        // TODO
+        return null;
+    }
+
+    public int GetDashboardsCount()
+    {
+        return GetDashboards().Count;
+    }
+}
