+++ Epam.ReportPortal.Automation/Epam.ReportPortal.Automation.ApiBusinessLayer/Epam.ReportPortal.Automation.ApiBusinessLayer.csproj
@@ -7,8 +7,7 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="log4net" Version="2.0.16" />
-    <PackageReference Include="Selenium.WebDriver" Version="4.22.0" />
+    <PackageReference Include="NUnit" Version="4.1.0" />
   </ItemGroup>
 
 </Project>
