+++ Epam.ReportPortal.Automation/Epam.ReportPortal.Automation.ApiTests/Base/ReportPortalApiTestsBase.cs
@@ -1,0 +1,18 @@
+﻿namespace Epam.ReportPortal.Automation.ApiTests.Base
+{
+    [TestFixture]
+    public class ReportPortalApiTestsBase
+    {
+        [SetUp]
+        public void BeforeEach()
+        {
+            // TODO
+        }
+
+        [TearDown]
+        public void AfterEach()
+        {
+            // TODO
+        }
+    }
+}
