+++ Epam.ReportPortal.Automation/Epam.ReportPortal.Automation.Configuration/Epam.ReportPortal.Automation.Configuration.csproj
@@ -1,4 +1,4 @@
-<Project Sdk="Microsoft.NET.Sdk">
+﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
     <TargetFramework>net6.0</TargetFramework>
@@ -7,8 +7,8 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="log4net" Version="2.0.16" />
-    <PackageReference Include="Selenium.WebDriver" Version="4.22.0" />
+    <Folder Include="Logger\" />
+    <Folder Include="Settings\" />
   </ItemGroup>
 
 </Project>
