--- Epam.ReportPortal.Automation/Epam.ReportPortal.Automation.Core/Selenium/Factory/IWebDriverFactory.cs
@@ -1,9 +1,0 @@
-﻿using OpenQA.Selenium;
-
-namespace Epam.ReportPortal.Automation.Core.Selenium.Factory
-{
-    public interface IWebDriverFactory
-    {
-        public IWebDriver? Create();
-    }
-}
