+++ Epam.ReportPortal.Automation/Epam.ReportPortal.Automation.CoreSelenium/Epam.ReportPortal.Automation.CoreSelenium.csproj
@@ -6,9 +6,4 @@
     <Nullable>enable</Nullable>
   </PropertyGroup>
 
-  <ItemGroup>
-    <PackageReference Include="log4net" Version="2.0.16" />
-    <PackageReference Include="Selenium.WebDriver" Version="4.22.0" />
-  </ItemGroup>
-
 </Project>
