+++ Epam.ReportPortal.Automation/Epam.ReportPortal.Automation.CoreSelenium/Selenium/Factory/ChromeDriverFactory.cs
@@ -1,0 +1,20 @@
+﻿namespace Epam.ReportPortal.Automation.CoreSelenium.Selenium.Factory
+{
+    public class ChromeDriverFactory: IWebDriverFactory 
+    {
+        public IWebDriver? Create()
+        {
+            var service = ChromeDriverService.CreateDefaultService("TODO");
+            var options = CreateDefaultChromeOptions();
+            var timeout = TimeSpan.FromSeconds(20);
+            var driver = new ChromeDriver(service, options, timeout);
+            return driver;
+        }
+
+        public static ChromeOptions CreateDefaultChromeOptions()
+        {
+            var options = new ChromeOptions();
+            return options;
+        }
+    }
+}
