+++ Epam.ReportPortal.Automation/Epam.ReportPortal.Automation.CoreSelenium/Selenium/Factory/IWebDriverFactory.cs
@@ -1,0 +1,7 @@
+﻿namespace Epam.ReportPortal.Automation.CoreSelenium.Selenium.Factory
+{
+    public interface IWebDriverFactory
+    {
+        public IWebDriver? Create();
+    }
+}
