--- Epam.ReportPortal.Automation/Epam.ReportPortal.Automation.Tests/ApiTests/Base/ReportPortalApiTestsBase.cs
@@ -1,20 +1,0 @@
-﻿using NUnit.Framework;
-
-namespace Epam.ReportPortal.Automation.Tests.ApiTests.Base
-{
-    [TestFixture]
-    internal class ReportPortalApiTestsBase
-    {
-        [OneTimeSetUp]
-        public void BeforeEach()
-        {
-            // TODO
-        }
-
-        [OneTimeTearDown]
-        public void AfterEach()
-        {
-            // TODO
-        }
-    }
-}
