+++ Epam.ReportPortal.Automation/Epam.ReportPortal.Automation.Tests/Epam.ReportPortal.Automation.Tests.csproj
@@ -7,10 +7,6 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <Folder Include="ApiTests\Dashboards\" />
-  </ItemGroup>
-
-  <ItemGroup>
     <PackageReference Include="log4net" Version="2.0.16" />
     <PackageReference Include="NUnit" Version="4.1.0" />
   </ItemGroup>
