+++ Epam.ReportPortal.Automation/Epam.ReportPortal.Automation.Tests/appSettings.json
@@ -4,4 +4,5 @@
     "Username": "testuser",
     "Password": "password"
   }
+
 }
