+++ Epam.ReportPortal.Automation/Epam.ReportPortal.Automation.UiTests/Base/ReportPortalUiTestsWithManyInstanesPerSuiteBase.cs
@@ -1,0 +1,20 @@
+﻿namespace Epam.ReportPortal.Automation.UiTests.Base
+{
+    [TestFixture]
+    public abstract class ReportPortalUiTestsWithManyInstancesPerSuiteBase
+    {
+        protected WebSteps WebSteps;
+
+        [OneTimeSetUp]
+        public void BeforeEach()
+        {
+            WebSteps = new WebSteps();
+        }
+
+        [OneTimeTearDown]
+        public void AfterEach()
+        {
+            // WebSteps.BrowserSteps.CloseBrowser();
+        }
+    }
+}
