+++ Epam.ReportPortal.Automation/Epam.ReportPortal.Automation.UiTests/Base/WebSteps.cs
@@ -1,0 +1,19 @@
+﻿namespace Epam.ReportPortal.Automation.UiTests.Base;
+
+public class WebSteps
+{
+    // private readonly Browser _browser;
+
+    public WebSteps()
+    {
+        // _browser = browser;
+    }
+
+    // public BrowserSteps BrowserSteps => new BrowserSteps(_browser);
+
+    // public LoginPageSteps LoginPageSteps => new LoginPageSteps(_browser);
+
+    // public AllDashboardsPageSteps AllDashboardsSteps => new AllDashboardsPageSteps(_browser);
+
+    // public ParticularDashboardPageSteps ParticularDashboardSteps => new ParticularDashboardPageSteps(_browser);
+}
