+++ Epam.ReportPortal.Automation/Epam.ReportPortal.Automation.UiTests/Common/GlobalUsings.cs
@@ -1,0 +1,1 @@
+global using NUnit.Framework;
