+++ Epam.ReportPortal.Automation/Epam.ReportPortal.Automation.UiTests/Dashboards/CreateDashboardsUiTests.cs
@@ -1,0 +1,8 @@
+﻿using Epam.ReportPortal.Automation.UiTests.Base;
+
+namespace Epam.ReportPortal.Automation.UiTests.Dashboards
+{
+    public class CreateDashboardsUiTests : ReportPortalUiTestsWithManyInstancesPerSuiteBase
+    {
+    }
+}
