+++ Epam.ReportPortal.Automation/Epam.ReportPortal.Automation.UiTests/Dashboards/RemoveDashboardsUiTests.cs
@@ -1,0 +1,8 @@
+﻿using Epam.ReportPortal.Automation.UiTests.Base;
+
+namespace Epam.ReportPortal.Automation.UiTests.Dashboards
+{
+    public class RemoveDashboardsUiTests : ReportPortalUiTestsWithManyInstancesPerSuiteBase
+    {
+    }
+}
