+++ Epam.ReportPortal.Automation/Epam.ReportPortal.Automation.UiTests/Epam.ReportPortal.Automation.UiTests.csproj
@@ -1,4 +1,4 @@
-<Project Sdk="Microsoft.NET.Sdk">
+﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
     <TargetFramework>net6.0</TargetFramework>
@@ -7,8 +7,7 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="log4net" Version="2.0.16" />
-    <PackageReference Include="Selenium.WebDriver" Version="4.22.0" />
+    <PackageReference Include="NUnit" Version="4.1.0" />
   </ItemGroup>
 
 </Project>
