+++ Epam.ReportPortal.Automation/Epam.ReportPortal.Automation.UiiBusinessLayer/WebObjects/BaseWebPage.cs
@@ -1,0 +1,12 @@
+﻿namespace Epam.ReportPortal.Automation.UiiBusinessLayer.WebObjects
+{
+    public abstract class BaseWebPage
+    {
+        protected readonly Browser Browser;
+
+        protected BaseWebPage(Browser browser)
+        {
+            Browser = browser;
+        }
+    }
+}
