+++ Epam.ReportPortal.Automation/Epam.ReportPortal.Automation.UiiBusinessLayer/WebSteps/Dashboards/ParticularDashboardPageSteps.cs
@@ -1,0 +1,31 @@
+﻿using Epam.ReportPortal.Automation.UiiBusinessLayer.WebObjects.Pages;
+
+namespace Epam.ReportPortal.Automation.UiiBusinessLayer.WebSteps.Dashboards
+{
+    public class ParticularDashboardPageSteps : BasePageSteps
+    {
+        private ParticularDashboardPage _particularDashboardPage;
+
+        public ParticularDashboardPageSteps(Browser browser) : base(browser)
+        {
+            _particularDashboardPage = new ParticularDashboardPage(browser);
+        }
+        
+        public void OpenParticularDashboardPage(string dashboardName)
+        {
+            // TODO
+        }
+
+        public List<string> GetWidgets()
+        {
+            // TODO
+            return null;
+        }
+
+        public int GetWidgetsCount()
+        {
+            // TODO
+            return -1;
+        }
+    }
+}
