+++ Epam.ReportPortal.Automation/Epam.ReportPortal.Automation.sln
@@ -3,11 +3,19 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
 VisualStudioVersion = 17.7.34031.279
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Epam.ReportPortal.Automation.Tests", "Epam.ReportPortal.Automation.Tests\Epam.ReportPortal.Automation.Tests.csproj", "{AAF5A2BA-CF5C-4AD9-A7CC-FE80DF3D8263}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Epam.ReportPortal.Automation.Core", "Epam.ReportPortal.Automation.Core\Epam.ReportPortal.Automation.Core.csproj", "{D97E82A8-F320-4792-ACDE-F4811F8BD409}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Epam.ReportPortal.Automation.BusinessLayer", "Epam.ReportPortal.Automation.BusinessLayer\Epam.ReportPortal.Automation.BusinessLayer.csproj", "{2FBEBFC2-A0AC-487D-B5A9-A8AB48240C6E}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Epam.ReportPortal.Automation.ApiTests", "Epam.ReportPortal.Automation.ApiTests\Epam.ReportPortal.Automation.ApiTests.csproj", "{C9187B13-90B3-4B98-B07F-898B4B5852D3}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Epam.ReportPortal.Automation.Core", "Epam.ReportPortal.Automation.Core\Epam.ReportPortal.Automation.Core.csproj", "{D97E82A8-F320-4792-ACDE-F4811F8BD409}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Epam.ReportPortal.Automation.UiTests", "Epam.ReportPortal.Automation.UiTests\Epam.ReportPortal.Automation.UiTests.csproj", "{2CE6F8C2-29E5-4E3A-B018-A71F0FEACE09}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Epam.ReportPortal.Automation.ApiBusinessLayer", "Epam.ReportPortal.Automation.ApiBusinessLayer\Epam.ReportPortal.Automation.ApiBusinessLayer.csproj", "{A77B41EC-F007-44E6-AE0F-3E57469D1765}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Epam.ReportPortal.Automation.UiiBusinessLayer", "Epam.ReportPortal.Automation.UiiBusinessLayer\Epam.ReportPortal.Automation.UiiBusinessLayer.csproj", "{D910F553-FF91-448B-AEFD-2C940C3C5DA6}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Epam.ReportPortal.Automation.Configuration", "Epam.ReportPortal.Automation.Configuration\Epam.ReportPortal.Automation.Configuration.csproj", "{E4254FD3-0B44-43B2-9C56-50DF2F103F4A}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Epam.ReportPortal.Automation.CoreSelenium", "Epam.ReportPortal.Automation.CoreSelenium\Epam.ReportPortal.Automation.CoreSelenium.csproj", "{8D4B8D33-4082-42A8-A05A-4CEF0AA778DF}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -15,18 +23,34 @@ Global
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{AAF5A2BA-CF5C-4AD9-A7CC-FE80DF3D8263}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{AAF5A2BA-CF5C-4AD9-A7CC-FE80DF3D8263}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{AAF5A2BA-CF5C-4AD9-A7CC-FE80DF3D8263}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{AAF5A2BA-CF5C-4AD9-A7CC-FE80DF3D8263}.Release|Any CPU.Build.0 = Release|Any CPU
-		{2FBEBFC2-A0AC-487D-B5A9-A8AB48240C6E}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{2FBEBFC2-A0AC-487D-B5A9-A8AB48240C6E}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{2FBEBFC2-A0AC-487D-B5A9-A8AB48240C6E}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{2FBEBFC2-A0AC-487D-B5A9-A8AB48240C6E}.Release|Any CPU.Build.0 = Release|Any CPU
 		{D97E82A8-F320-4792-ACDE-F4811F8BD409}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{D97E82A8-F320-4792-ACDE-F4811F8BD409}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{D97E82A8-F320-4792-ACDE-F4811F8BD409}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{D97E82A8-F320-4792-ACDE-F4811F8BD409}.Release|Any CPU.Build.0 = Release|Any CPU
+		{C9187B13-90B3-4B98-B07F-898B4B5852D3}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{C9187B13-90B3-4B98-B07F-898B4B5852D3}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{C9187B13-90B3-4B98-B07F-898B4B5852D3}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{C9187B13-90B3-4B98-B07F-898B4B5852D3}.Release|Any CPU.Build.0 = Release|Any CPU
+		{2CE6F8C2-29E5-4E3A-B018-A71F0FEACE09}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{2CE6F8C2-29E5-4E3A-B018-A71F0FEACE09}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{2CE6F8C2-29E5-4E3A-B018-A71F0FEACE09}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{2CE6F8C2-29E5-4E3A-B018-A71F0FEACE09}.Release|Any CPU.Build.0 = Release|Any CPU
+		{A77B41EC-F007-44E6-AE0F-3E57469D1765}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{A77B41EC-F007-44E6-AE0F-3E57469D1765}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{A77B41EC-F007-44E6-AE0F-3E57469D1765}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{A77B41EC-F007-44E6-AE0F-3E57469D1765}.Release|Any CPU.Build.0 = Release|Any CPU
+		{D910F553-FF91-448B-AEFD-2C940C3C5DA6}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{D910F553-FF91-448B-AEFD-2C940C3C5DA6}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{D910F553-FF91-448B-AEFD-2C940C3C5DA6}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{D910F553-FF91-448B-AEFD-2C940C3C5DA6}.Release|Any CPU.Build.0 = Release|Any CPU
+		{E4254FD3-0B44-43B2-9C56-50DF2F103F4A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{E4254FD3-0B44-43B2-9C56-50DF2F103F4A}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{E4254FD3-0B44-43B2-9C56-50DF2F103F4A}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{E4254FD3-0B44-43B2-9C56-50DF2F103F4A}.Release|Any CPU.Build.0 = Release|Any CPU
+		{8D4B8D33-4082-42A8-A05A-4CEF0AA778DF}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{8D4B8D33-4082-42A8-A05A-4CEF0AA778DF}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{8D4B8D33-4082-42A8-A05A-4CEF0AA778DF}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{8D4B8D33-4082-42A8-A05A-4CEF0AA778DF}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
