+++ Epam.ReportPortal.Automation/Epam.ReportPortal.AutomationSelenium/Class1.cs
@@ -1,0 +1,7 @@
+﻿namespace Epam.ReportPortal.AutomationSelenium
+{
+    public class Class1
+    {
+
+    }
+}
